monitor/requirements.txt
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
azure-core==1.26.4
azure-cosmos==4.3.1
azure-identity==1.12.0
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
Flask==2.3.1
Flask-Cors==3.0.10
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
msal==1.22.0
msal-extensions==1.0.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
portalocker==2.7.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
requests==2.29.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.2
yarl==1.9.2
